+++ CHANGELOG.md
@@ -2,5 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.1] - 2024-08-27
+### Changed
+- ADDON_VIEWS_COUNT moved to property from const
+### Fixed
+- Update the scroll position by a scrollbar
+
 ## [2.0.0] - 2024-04-02
 ### Release v2